+++ src/ffi/cpython/mod.rs
@@ -1,3 +1,11 @@
 pub mod abstract_;
+#[cfg(not(PyPy))]
+pub mod bytesobject;
+pub mod ceval;
+pub mod code;
 
 pub use self::abstract_::*;
+#[cfg(not(PyPy))]
+pub use self::bytesobject::*;
+pub use self::ceval::*;
+pub use self::code::*;
